+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Aggregating namespace message "Initializing empty JSON"
 - Default logger name as "🌐Translate"
+- translate function, options parameter validation
 
 
 ---